requirements.txt
# ============================================
# KrishiMitra - Agricultural AI ChatBot
# Python Dependencies
# ============================================

# ==============================================
# Core Framework
# ==============================================
fastapi>=0.124.0
uvicorn[standard]>=0.38.0
python-multipart>=0.0.21
python-dotenv>=1.1.0
pydantic>=2.12.0
pydantic-settings>=2.12.0

# ==============================================
# LangChain & RAG
# ==============================================
langchain>=0.2.14
langchain-community>=0.2.12
langchain-core>=1.2.0
langchain-chroma>=1.1.0
langchain-huggingface>=1.2.0
langchain-openai>=0.1.20
langchain-text-splitters>=0.2.4

# ==============================================
# Vector Databases
# ==============================================
chromadb>=1.3.0
faiss-cpu>=1.13.0

# ==============================================
# Embeddings & ML
# ==============================================
sentence-transformers>=5.2.0
transformers>=4.57.0
huggingface-hub>=0.36.0
torch>=2.9.0
accelerate>=1.12.0
safetensors>=0.7.0

# ==============================================
# LLM Client
# ==============================================
groq>=0.37.0
openai>=1.109.0
tiktoken>=0.12.0

# ==============================================
# Data Processing
# ==============================================
numpy>=1.26.0
pandas>=2.3.0
scikit-learn>=1.7.0
scipy>=1.15.0
joblib>=1.5.0

# ==============================================
# HTTP & Networking
# ==============================================
requests>=2.32.0
httpx>=0.28.0
aiohttp>=3.13.0

# ==============================================
# Web Scraping (for mandi prices)
# ==============================================
beautifulsoup4>=4.14.0

# ==============================================
# Database & ORM
# ==============================================
SQLAlchemy>=2.0.0

# ==============================================
# Utilities
# ==============================================
PyYAML>=6.0.0
python-dateutil>=2.9.0
tqdm>=4.67.0
rich>=14.2.0
colorama>=0.4.6
regex>=2025.11.0

# ==============================================
# Monitoring & Telemetry (ChromaDB dependencies)
# ==============================================
opentelemetry-api>=1.39.0
opentelemetry-sdk>=1.39.0
opentelemetry-instrumentation-fastapi>=0.60b1

# ==============================================
# Additional Dependencies
# ==============================================
orjson>=3.11.0
protobuf>=6.33.0
onnxruntime>=1.23.0
